+++ examples/addons/log-events.py
@@ -1,8 +1,11 @@
 """Post messages to mitmproxy's event log."""
-from mitmproxy import ctx
+import logging
+
+from mitmproxy.log import ALERT
 
 
 def load(l):
-    ctx.log.info("This is some informative text.")
-    ctx.log.warn("This is a warning.")
-    ctx.log.error("This is an error.")
+    logging.info("This is some informative text.")
+    logging.warning("This is a warning.")
+    logging.error("This is an error.")
+    logging.log(ALERT, "This is an alert. It has the same urgency as info, but will also pop up in the status bar.")