+++ test/mitmproxy/proxy/conftest.py
@@ -5,7 +5,6 @@ from hypothesis import settings
 
 from mitmproxy import connection, options
 from mitmproxy.addons.proxyserver import Proxyserver
-from mitmproxy.addons.termlog import TermLog
 from mitmproxy.proxy import context
 
 
@@ -13,7 +12,6 @@ from mitmproxy.proxy import context
 def tctx() -> context.Context:
     opts = options.Options()
     Proxyserver().load(opts)
-    TermLog().load(opts)
     return context.Context(
         connection.Client(("client", 1234), ("127.0.0.1", 8080), 1605699329), opts
     )